\name{getPWM}
\alias{getPWM}
\alias{getPWM-methods}
\alias{getPWM,motiv-method}
\title{Get PWMs from a motiv object} 

\description{
Get PWMs from a motiv object.} 
\usage{
	\S4method{getPWM}{motiv}(x)
}
\arguments{
	\item{x}{An object of class \code{motiv}.}
}

\author{Eloi Mercier <\email{emercier@chibi.ubc.ca}>}

\keyword{misc}
 
